PROD: Daily Message sender not setup Fixes #394
